Continue refactoring under HAL
